Nix : Add initial support nix develop
Initial work in supporting Nix packaging. It introduces
a devshell for getting started in development on Nix based environments.

In addition an direnv is defined to make it easy to transition into the
development environment. There are 3 supported ways:

With Flakes enabled on your host
`nix develop`

Without flakes support
`nix-shell`

With direnv enabled
`direnv allow`

And then build in the normal way.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>